cleanup in tls data

+++ drace-client/include/drace-client.h
@@ -75,7 +75,7 @@ extern int      tls_idx;
 // TODO check if global is better
 static std::atomic<int> num_threads_active{ 0 };
 extern std::atomic<uint> runtime_tid;
-extern std::atomic<uint64> last_th_start;
+extern std::atomic<uint> last_th_start;
 extern std::atomic<bool> th_start_pending;
 
 // Module
